% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\docType{data}
\name{app}
\alias{app}
\title{Example data: mnem results for
the Dixit et al., 2016 and Datlinger et al., pooled CRISPR screens.
For details see the vignette.}
\description{
Example data: mnem results for
the Dixit et al., 2016 and Datlinger et al., pooled CRISPR screens.
For details see the vignette.
}
\examples{
data(app)
}
