% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\name{bootstrap}
\alias{bootstrap}
\title{run bootstrap simulations on the components}
\usage{
bootstrap(x, size = 1000, logtype = 2, ...)
}
\arguments{
\item{x}{mnem object}

\item{size}{size of the booststrap simulations}

\item{logtype}{logarithm type of the data}

\item{...}{additional parameters for hte nem function}
}
\value{
returns bootstrap support for each edge
}
\description{
run bootstrap simulations on the components
}
\examples{
sim <- simData(Sgenes = 3, Egenes = 2, Nems = 2, mw = c(0.4,0.6))
data <- (sim$data - 0.5)/0.5
data <- data + rnorm(length(data), 0, 1)
result <- mnem(data, k = 2, starts = 2)
boot <- bootstrap(result, size = 10)
}
\author{
Martin Pirkl
}
