% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\name{getAffinity}
\alias{getAffinity}
\title{Calculate responsibilities}
\usage{
getAffinity(x, affinity = 0, norm = TRUE, logtype = 2, mw = NULL,
  data = matrix(0, 2, ncol(x)))
}
\arguments{
\item{x}{log odds for l cells and k components as a kxl matrix}

\item{logtype}{logarithm type of the data}

\item{mw}{miture weights}

\item{data}{data in log odds}
}
\description{
Calculate responsibilities
}
\examples{
#' sim <- simData(Sgenes = 3, Egenes = 2, Nems = 2, mw = c(0.4,0.6))
data <- (sim$data - 0.5)/0.5
data <- data + rnorm(length(data), 0, 1)
result <- mnem(data, k = 2, starts = 2)
resp <- getAffinity(result$probs, mw = result$mw, data = data)
}
\author{
Martin Pirkl
}
