% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\name{getIC}
\alias{getIC}
\title{plot mnem result}
\usage{
getIC(x, man = FALSE, degree = 4, logtype = 2, pen = 2, useF = FALSE,
  Fnorm = FALSE)
}
\arguments{
\item{x}{mnem object}

\item{man}{logical. manual data penalty}

\item{degree}{different degree of complexity}

\item{logtype}{logarithm type of the data}

\item{pen}{penalty weight ofr the data}

\item{useF}{use F (see publication) as complexity}

\item{Fnorm}{normalize complexity of F}
}
\value{
penlized likelihood
}
\description{
plot mnem result
}
\examples{
sim <- simData(Sgenes = 5, Egenes = 2, Nems = 2, mw = c(0.4,0.6))
data <- (sim$data - 0.5)/0.5
data <- data + rnorm(length(data), 0, 1)
pen <- numeric(3)
result <- list()
for (k in 1:3) {
result[[k]] <- mnem(data, k = k, starts = 10)
pen[k] <- getIC(result[[k]])
}
print(pen)
}
\author{
Martin Pirkl
}
