% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\name{hamSim}
\alias{hamSim}
\title{normalized hamming accuracy of the columns of two adjacency matrices}
\usage{
hamSim(a, b, diag = 1, symmetric = TRUE)
}
\arguments{
\item{a}{adjacency matrix}

\item{b}{adjacency matrix}

\item{diag}{if 1 includes diagonal in distance, if 0 not}

\item{symmetric}{comparing a to b is asymmetrical, if TRUE includes
comparison b to a}
}
\value{
normalized hamming accuracy
}
\description{
normalized hamming accuracy of the columns of two adjacency matrices
}
\examples{
sim <- simData(Sgenes = 3, Egenes = 2, Nems = 2, mw = c(0.4,0.6))
data <- (sim$data - 0.5)/0.5
data <- data + rnorm(length(data), 0, 1)
result <- mnem(data, k = 2, starts = 2)
plot(result)
}
\author{
Martin Pirkl
}
