% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\name{mnem}
\alias{mnem}
\title{learn mixture of networks from a single-ceel knock-down experiment}
\usage{
mnem(D, inference = "em", search = "modules", start = NULL,
  method = "llr", parallel = NULL, reduce = FALSE, runs = 1,
  starts = 3, type = "random", p = NULL, k = NULL, kmax = 10,
  verbose = FALSE, max_iter = 100, parallel2 = NULL, converged = 10^-1,
  redSpace = NULL, affinity = 0, evolution = FALSE, subtopoX = NULL,
  ratio = TRUE, logtype = 2, initnets = FALSE, popSize = 10,
  stallMax = 2, elitism = NULL, maxGens = Inf, domean = TRUE,
  modulesize = 5, compress = FALSE)
}
\arguments{
\item{D}{data with cells indexing the columns and features (E-genes) indexing the rows}

\item{inference}{inference method "em" for expectation maximization, "greedy" or "genetic"}

\item{search}{search method for single network inference "greedy", "exhaustive" or "modules"}

\item{start}{A list of n lists of k networks for n starts of the EM and k components}

\item{method}{"llr" for log ratios or foldchanges as input (see ratio)}

\item{parallel}{number of threads}

\item{reduce}{logical - reduce search space for exhaustive search}

\item{runs}{how many runs for greedy search}

\item{starts}{how many starts for em}

\item{type}{initialize with "random" probabilities or "cluster" the data}

\item{p}{initial probabilities}

\item{k}{number of components}

\item{kmax}{maximum number of components when k=NULL is inferred}

\item{verbose}{output}

\item{max_iter}{maximum iteration if likelihood does not converge}

\item{parallel2}{if parallel=NULL, number of threads for component optimization}

\item{converged}{absolute distance for convergence}

\item{redSpace}{space for exhaustive search}

\item{affinity}{0 is default, 1 is for hard clustering}

\item{evolution}{logical, penalty on difference of components}

\item{subtopoX}{hard prior on theta}

\item{ratio}{logical, if true data is log ratios, if false foldchanges}

\item{logtype}{logarithm of the data}

\item{initnets}{initial networks in a list}

\item{popSize}{population size}

\item{stallMax}{maximum number of stall generations till convergence}

\item{elitism}{number of networks to keep during evolution}

\item{maxGens}{maximum numbero f generations}

\item{domean}{average the data (speed imporvment)}

\item{compress}{compress networks after search}

\item{modulsize}{max number of S-genes per module in module search}
}
\value{
optimized network for data fit
}
\description{
learn mixture of networks from a single-ceel knock-down experiment
}
\examples{
sim <- simData(Sgenes = 5, Egenes = 2, Nems = 2, mw = c(0.4,0.6))
data <- (sim$data - 0.5)/0.5
data <- data + rnorm(length(data), 0, 1)
result <- mnem(data, k = 2, starts = 10)
plot(result)
}
\author{
Martin Pirkl
}
