% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\name{plot.mnemsim}
\alias{plot.mnemsim}
\title{Plot simulated mixture.}
\usage{
\method{plot}{mnemsim}(x, ...)
}
\arguments{
\item{x}{mnemsim object}

\item{...}{additional parameters for the plotting function plotDNF}
}
\value{
visualization of simulated mixture with Rgraphviz
}
\description{
Plot simulated mixture.
}
\examples{
sim <- simData(Sgenes = 3, Egenes = 2, Nems = 2, mw = c(0.4,0.6))
plot(sim)
}
\author{
Martin Pirkl
}
