% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\name{transitive.reduction}
\alias{transitive.reduction}
\title{Transitive reduction}
\usage{
transitive.reduction(g)
}
\arguments{
\item{g}{adjacency matrix or graphNEL object}
}
\value{
transitively reduced adjacency matrix
}
\description{
Computes the transitive reduction of an adjacency
matrix or graphNEL object.
Originally imported from the package 'nem'.
}
\examples{
g <- matrix(c(0,0,0,1,0,0,0,1,0), 3)
rownames(g) <- colnames(g) <- seq_len(3)
g.tr <- transitive.reduction(g)
}
\references{
R. Sedgewick, Algorithms, Pearson, 2002.
}
\author{
Holger Froehlich
}
