% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\docType{data}
\name{app}
\alias{app}
\title{Processed scRNAseq from pooled CRISPR screens
Example data: mnem results for
the Dixit et al., 2016 and Datlinger et al., pooled CRISPR screens.
For details see the vignette.}
\usage{
app
}
\description{
Processed scRNAseq from pooled CRISPR screens
Example data: mnem results for
the Dixit et al., 2016 and Datlinger et al., pooled CRISPR screens.
For details see the vignette.
}
\examples{
data(app)
}
\references{
Datlinger, P., Rendeiro, A., Schmidl, C., Krausgruber, T.,
Traxler, P., Klughammer, J., Schuster, L. C., Kuchler, A., Alpar, D.,
and Bock, C. (2017). Pooled crispr screening with single-cell transcriptome
readout. Nature Methods, 14, 297 EP ???.

Dixit, A., Parnas, O., Li, B., Chen, J., Fulco, C. P.,
Jerby-Arnon, L., Marjanovic, N. D., Dionne, D., Burks, T., Raychowdhury, R.,
Adamson, B., Norman, T. M., Lander, E. S., Weissman, J. S., Friedman, N., and
Regev, A. (2016). Perturb-seq: Dissecting molecular circuits with scalable
single-cell rna profiling of pooled genetic screens. Cell, 167(7),
1853???1866.e17.
}
