% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\name{clustNEM}
\alias{clustNEM}
\title{Cluster NEM.}
\usage{
clustNEM(data, k = 2:5, cluster = NULL, learnkpars = list(), ...)
}
\arguments{
\item{data}{data of log ratios with cells in columns and features in rows}

\item{k}{number of clusters}

\item{cluster}{given clustering has to correspond to the columns of data}

\item{learnkpars}{list of additional arguments for functio learnk, if
cluster is NULL}

\item{...}{additional arguments for standard nem function}
}
\value{
family of nems; the first k list entries hold full information of
the standard nem search
\item{comp}{list of all adjacency matrices phi}
\item{mw}{vector of mixture weights}
\item{probs}{fake cell probabilities (see mw: mixture weights)}
}
\description{
This function clusters the data and performs standard nem on each cluster.
}
\examples{
sim <- simData(Sgenes = 3, Egenes = 2, Nems = 2, mw = c(0.4,0.6))
data <- (sim$data - 0.5)/0.5
data <- data + rnorm(length(data), 0, 1)
resulst <- clustNEM(data, k = 2:3)
}
\author{
Martin Pirkl
}
