% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\name{fuzzyindex}
\alias{fuzzyindex}
\title{Calculate fuzzy ground truth.}
\usage{
fuzzyindex(x, data, logtype = 2, ...)
}
\arguments{
\item{x}{mnemsim object}

\item{data}{noisy data matrix}

\item{logtype}{logarithm type of the data}

\item{...}{additional parameters for the function getAffinity}
}
\value{
list with cell log odds mixture weights and log likelihood
}
\description{
Calculates responsibilities and mixture weights based on
the ground truth and noisy data.
}
\examples{
sim <- simData(Sgenes = 3, Egenes = 2, Nems = 2, mw = c(0.4,0.6))
data <- sim$data
data[which(sim$data == 1)] <- rnorm(sum(sim$data == 1), 1, 1)
data[which(sim$data == 0)] <- rnorm(sum(sim$data == 0), -1, 1)
fuzzy <- fuzzyindex(sim, data)
}
\author{
Martin Pirkl
}
