% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\name{plot.mnem}
\alias{plot.mnem}
\title{plot mnem result}
\usage{
plot.mnem(x, oma = c(3, 1, 1, 3), main = "M&NEM", anno = TRUE,
  cexAnno = 1, scale = NULL, global = TRUE, egenes = TRUE,
  sep = FALSE, tsne = FALSE, affinity = 0, logtype = 2, cells = TRUE,
  pch = ".", legend = FALSE, showdata = FALSE, bestCell = TRUE,
  showprobs = FALSE, shownull = TRUE, ratio = TRUE, method = "llr",
  showweights = TRUE, ...)
}
\arguments{
\item{x}{mnem object}

\item{oma}{outer margin}

\item{main}{main text}

\item{anno}{annotate cells}

\item{cexAnno}{text size}

\item{scale}{scale cells}

\item{global}{global clustering}

\item{egenes}{show egene attachments}

\item{sep}{seperate clusters}

\item{tsne}{use tsne}

\item{affinity}{logical use hard clustering if true}

\item{logtype}{logarithm type of the data}

\item{cells}{show ccell attachments}

\item{pch}{cell symbol}

\item{legend}{show legend}

\item{showdata}{show data if true}

\item{bestCell}{show probability of best fitting cell}

\item{showprobs}{show probabilities}

\item{shownull}{show null node}

\item{ratio}{use log ratios (if true) or foldchanges}

\item{method}{"llr" for ratios}

\item{showweights}{shwo weights}

\item{...}{additional parameters}
}
\value{
visualization of mnem result Rgraphviz
}
\description{
plot mnem result
}
\examples{
sim <- simData(Sgenes = 3, Egenes = 2, Nems = 2, mw = c(0.4,0.6))
data <- (sim$data - 0.5)/0.5
data <- data + rnorm(length(data), 0, 1)
result <- mnem(data, k = 2, starts = 2)
plot(result)
}
\author{
Martin Pirkl
}
