% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\docType{data}
\name{sim}
\alias{sim}
\title{Example data: simulation results
Contains simulation results. How they were
aquired is explained in the vignette.
The data conists of a list of data matrices holding
sensitivity and specificity
(spec, sens) of network edges for the variious methods compared to
the ground truth, sensitivity and specificity (sens2, spec2)
of the expected data for clustNEM and mnem.}
\description{
Example data: simulation results
Contains simulation results. How they were
aquired is explained in the vignette.
The data conists of a list of data matrices holding
sensitivity and specificity
(spec, sens) of network edges for the variious methods compared to
the ground truth, sensitivity and specificity (sens2, spec2)
of the expected data for clustNEM and mnem.
}
\examples{
data(sim)
}
