% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\docType{data}
\name{sim}
\alias{sim}
\title{Simulation results
Example data: simulation results
Contains four lists with results of mixtures with 3, 5, 10 and 20 S-genes.
Each list consists of an array including 100 runs for three different
noise levels, five different numbers of components (1,2,3,4,5) and three
different methods, which results in overall 2400 runs. For each run the
array includes values for time, over-fit, component accuracy (accuracy) and
mixing weight accuracy (mixing).}
\usage{
sim
}
\description{
Simulation results
Example data: simulation results
Contains four lists with results of mixtures with 3, 5, 10 and 20 S-genes.
Each list consists of an array including 100 runs for three different
noise levels, five different numbers of components (1,2,3,4,5) and three
different methods, which results in overall 2400 runs. For each run the
array includes values for time, over-fit, component accuracy (accuracy) and
mixing weight accuracy (mixing).
}
\examples{
data(sim)
}
