% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems.r
\docType{data}
\name{sim}
\alias{sim}
\title{Simulation results
Example data: simulation results
Contains simulation results.
For details see the vignette.}
\usage{
sim
}
\description{
Simulation results
Example data: simulation results
Contains simulation results.
For details see the vignette.
}
\examples{
data(sim)
}
