% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnems_low.r
\name{simData}
\alias{simData}
\title{simulate single cell data from a mixture of networks}
\usage{
simData(Sgenes = 5, Egenes = 1, subsample = 1, Nems = 2, reps = NULL,
  mw = NULL, evolution = FALSE, nCells = 1000, uninform = 0,
  unitheta = FALSE)
}
\arguments{
\item{Sgenes}{number of Sgenes}

\item{Egenes}{number of Egenes}

\item{subsample}{range to subsample data. 1 means the full simulated data is used.}

\item{Nems}{numberof components}

\item{reps}{number of relicates, if set (not realistice for cells)}

\item{mw}{mixture weights}

\item{evolution}{evovling network, if set to true}

\item{nCells}{number of cells}

\item{uninform}{number of uninformative Egenes}

\item{unitheta}{uniform theta, if true}
}
\value{
simulation object with meta information and data
}
\description{
simulate single cell data from a mixture of networks
}
\examples{
sim <- simData(Sgenes = 5, Egenes = 2, Nems = 3, mw = c(0.1,0.3,0.6))
data <- sim$data
data <- data + rnorm(length(data), 0, 0.5)
result <- mnem(data, k = 3)
plot(result)
}
\author{
Martin Pirkl
}
